[AJ-1296] Organize data tables UI part 1 (#4162)
