+++ src/components/data/LocalVariablesContent.js
@@ -4,12 +4,12 @@ import { Fragment, useEffect, useState } from 'react';
 import { div, h } from 'react-hyperscript-helpers';
 import { AutoSizer } from 'react-virtualized';
 import { DeleteConfirmationModal, Link, Select, spinnerOverlay } from 'src/components/common';
-import { renderDataCell } from 'src/components/data/data-utils';
 import Dropzone from 'src/components/Dropzone';
 import FloatingActionButton from 'src/components/FloatingActionButton';
 import { icon } from 'src/components/icons';
 import { DelayedSearchInput, TextInput } from 'src/components/input';
 import { FlexTable, HeaderCell } from 'src/components/table';
+import { renderDataCell } from 'src/data/data-table/entity-service/renderDataCell';
 import { Ajax } from 'src/libs/ajax';
 import colors from 'src/libs/colors';
 import { withErrorReporting } from 'src/libs/error';
