+++ src/components/data/UploadPreviewTable.js
@@ -3,11 +3,11 @@ import { Fragment, useEffect, useMemo, useRef, useState } from 'react';
 import { code, div, em, h, h3, p, span, strong } from 'react-hyperscript-helpers';
 import { AutoSizer } from 'react-virtualized';
 import { ButtonPrimary, ButtonSecondary, fixedSpinnerOverlay } from 'src/components/common';
-import { renderDataCell } from 'src/components/data/data-utils';
 import { icon } from 'src/components/icons';
 import { NameModal } from 'src/components/NameModal';
 import { GridTable, HeaderCell, Resizable, Sortable } from 'src/components/table';
 import TooltipTrigger from 'src/components/TooltipTrigger';
+import { renderDataCell } from 'src/data/data-table/entity-service/renderDataCell';
 import { Ajax } from 'src/libs/ajax';
 import colors from 'src/libs/colors';
 import { withErrorReporting } from 'src/libs/error';
